chore: work on code quality of extensions

+++ src/TimeofFlight.cpp
@@ -1,30 +1,23 @@
 #include "TimeofFlight.h"
+
 #include "Adafruit_VL53L0X.h"
 
-bool TimeOfFlight::begin()
-{
+bool TimeOfFlight::begin() {
     if (!sensor.begin()) {
         return false;
     }
     return true;
 }
 
-int TimeOfFlight::read()
-{
+int TimeOfFlight::read() {
     VL53L0X_RangingMeasurementData_t measure;
     sensor.rangingTest(&measure, false);
-    if (measure.RangeStatus == 4) { 
+    if (measure.RangeStatus == 4) {
         return -1;
     }
-    return measure.RangeMilliMeter;
+    return static_cast<int>(measure.RangeMilliMeter);
 }
 
-int TimeOfFlight::readCm()
-{
-    return this->read() / 10;
-}
+int TimeOfFlight::readCm() { return this->read() / 10; }
 
-int TimeOfFlight::readMm()
-{
-    return this->read();
-}
+int TimeOfFlight::readMm() { return this->read(); }
